+++ src/ScriptUnit.cpp
@@ -227,7 +227,7 @@ void ScriptUnit::removeScript( TScript * pT )
 }
 
 
-qint64 ScriptUnit::getNewID()
+int ScriptUnit::getNewID()
 {
     return ++mMaxID;
 }
